chore(ECO-3169): Organize arena components into separate files
